tests: move mocks definition to unit_bin target
Mocked functions are only used for unit tests, but they are defined in
the harness target, which is only linked to the core module. Hence,
functions defined in src/bpfilter can't be mocked.

Move the mocks definition to the unit tests target, and mock
bf_ctx_token() to ensure maps can be tested properly.
