+++ tests/unit/bpfilter/cgen/cgen.c
@@ -7,7 +7,7 @@
 
 #include "fake.h"
 #include "harness/test.h"
-#include "harness/mock.h"
+#include "mock.h"
 
 Test(cgen, create_delete_assert)
 {
