+++ tests/unit/bpfilter/cgen/printer.c
@@ -7,7 +7,7 @@
 
 #include "fake.h"
 #include "harness/test.h"
-#include "harness/mock.h"
+#include "mock.h"
 
 Test(printer, msg_lifetime)
 {
