+++ tests/unit/bpfilter/cgen/prog/map.c
@@ -7,7 +7,7 @@
 
 #include "fake.h"
 #include "harness/test.h"
-#include "harness/mock.h"
+#include "mock.h"
 
 Test(map, create_delete_assert)
 {
@@ -141,7 +141,8 @@ Test(map, map_create_assert)
 Test(map, map_create)
 {
     _cleanup_bf_map_ struct bf_map *map = NULL;
-    _clean_bf_test_mock_ bf_test_mock _ = bf_test_mock_get(bf_bpf, 16);
+    _clean_bf_test_mock_ bf_test_mock _0 = bf_test_mock_get(bf_bpf, 16);
+    _clean_bf_test_mock_ bf_test_mock _1 = bf_test_mock_get(bf_ctx_token, -1);
 
     assert_success(bf_map_new(&map, "suffix", BF_MAP_TYPE_SET, BF_MAP_BPF_TYPE_ARRAY, 1, 1, BF_MAP_N_ELEMS_UNKNOWN));
     assert_error(bf_map_create(map, 0));
@@ -156,7 +157,8 @@ Test(map, map_create)
 Test(map, map_create_failure)
 {
     _cleanup_bf_map_ struct bf_map *map = NULL;
-    _clean_bf_test_mock_ bf_test_mock _ = bf_test_mock_get(bf_bpf, -1);
+    _clean_bf_test_mock_ bf_test_mock _0 = bf_test_mock_get(bf_bpf, -1);
+    _clean_bf_test_mock_ bf_test_mock _1 = bf_test_mock_get(bf_ctx_token, -1);
 
     assert_success(bf_map_new(&map, "suffix", BF_MAP_TYPE_SET, BF_MAP_BPF_TYPE_ARRAY, 1, 1, 1));
     assert_error(bf_map_create(map, 0));
