+++ tests/unit/bpfilter/cgen/program.c
@@ -7,7 +7,7 @@
 
 #include "fake.h"
 #include "harness/test.h"
-#include "harness/mock.h"
+#include "mock.h"
 
 Test(program, emit_fixup_call)
 {
