+++ tests/unit/bpfilter/cgen/swich.c
@@ -7,7 +7,7 @@
 
 #include "fake.h"
 #include "harness/test.h"
-#include "harness/mock.h"
+#include "mock.h"
 
 Test(swich, new_and_free_option)
 {
