+++ tests/unit/bpfilter/xlate/nft/nfgroup.c
@@ -9,7 +9,7 @@
 
 #include "fake.h"
 #include "harness/test.h"
-#include "harness/mock.h"
+#include "mock.h"
 
 Test(nfgroup, new_and_free)
 {
