+++ tests/unit/bpfilter/xlate/nft/nfmsg.c
@@ -9,7 +9,7 @@
 
 #include "fake.h"
 #include "harness/test.h"
-#include "harness/mock.h"
+#include "mock.h"
 
 Test(nfmsg, new_and_free)
 {
