+++ tests/unit/core/marsh.c
@@ -6,7 +6,7 @@
 #include "core/marsh.c"
 
 #include "harness/test.h"
-#include "harness/mock.h"
+#include "mock.h"
 
 Test(marsh, new)
 {
