+++ tests/unit/core/matcher.c
@@ -6,7 +6,7 @@
 #include "core/matcher.c"
 
 #include "harness/test.h"
-#include "harness/mock.h"
+#include "mock.h"
 
 Test(matcher, new_and_free)
 {
